Adding definition for Ubuntu 10.04

+++ definitions/.ubuntu/cleanup.sh
@@ -1,8 +1,11 @@
+#!/bin/bash -eux
 # Remove items used for building, since they aren't needed anymore
 
 #Clean up tmp
 rm -rf /tmp/*
 
-# Removing leftover leases and persistent rules
-echo "cleaning up dhcp leases"
-rm /var/lib/dhcp/*
+if [ -d "/var/lib/dhcp" ]; then
+    # Removing leftover leases and persistent rules
+    echo "cleaning up dhcp leases"
+    rm /var/lib/dhcp/*
+fi
